+++ src/main/java/tk/themcbros/uselessmod/proxy/CommonProxy.java
@@ -1,5 +1,6 @@
 package tk.themcbros.uselessmod.proxy;
 
+import net.minecraft.item.ItemStack;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.fml.InterModComms;
 import net.minecraftforge.fml.ModList;
@@ -21,6 +22,7 @@ import tk.themcbros.uselessmod.handler.ForgeEventHandlers;
 import tk.themcbros.uselessmod.handler.UselessPlayerEvents;
 import tk.themcbros.uselessmod.lists.ModBiomes;
 import tk.themcbros.uselessmod.lists.ModEntityTypes;
+import tk.themcbros.uselessmod.lists.ModItems;
 import tk.themcbros.uselessmod.lists.VanillaCompat;
 import tk.themcbros.uselessmod.world.OreGeneration;
 
@@ -47,15 +49,15 @@ public class CommonProxy {
 
 		VanillaCompat.register();
 
-//		if (ModList.get().isLoaded("bluepower")) {
-//			UselessMod.LOGGER.info("Registering recycling recipes for BluePower");
-//			com.bluepowermod.api.recipe.IAlloyFurnaceRegistry furnaceRegistry = com.bluepowermod.api.BPApi.getInstance().getAlloyFurnaceRegistry();
-//			furnaceRegistry.addRecyclingRecipe(new ItemStack(ModItems.USELESS_INGOT));
-//			furnaceRegistry.addRecyclingRecipe(new ItemStack(ModItems.SUPER_USELESS_INGOT));
-//			furnaceRegistry.addRecyclingRecipe(new ItemStack(ModItems.USELESS_NUGGET));
-//			furnaceRegistry.addRecyclingRecipe(new ItemStack(ModItems.SUPER_USELESS_NUGGET));
-//		}
-//
+		if (ModList.get().isLoaded("bluepower")) {
+			UselessMod.LOGGER.info("Registering recycling recipes for BluePower");
+			com.bluepowermod.api.recipe.IAlloyFurnaceRegistry furnaceRegistry = com.bluepowermod.api.BPApi.getInstance().getAlloyFurnaceRegistry();
+			furnaceRegistry.addRecyclingRecipe(new ItemStack(ModItems.USELESS_INGOT));
+			furnaceRegistry.addRecyclingRecipe(new ItemStack(ModItems.SUPER_USELESS_INGOT));
+			furnaceRegistry.addRecyclingRecipe(new ItemStack(ModItems.USELESS_NUGGET));
+			furnaceRegistry.addRecyclingRecipe(new ItemStack(ModItems.SUPER_USELESS_NUGGET));
+		}
+
 //		if (ModList.get().isLoaded("ironjetpacks")) {
 //			Jetpack uselessJet = JetpackRegistry.createJetpack("useless", 3, 0xfff, 9, 30, "tag:forge:ingots/useless");
 //			Jetpack superUselessJet = JetpackRegistry.createJetpack("super_useless", 4, 0xfff, 9, 30, "tag:forge:ingots/super_useless");